requirements.txt
requests
python-dotenv
streamlit
pandas
numpy
scikit-learn
shapely
# sqlite3 já incluso no Python padrão
